THE GENDER COUNTING DATABASE
Ever been given the responsibility of counting the amount of individual that comes into an event, or a party and you loose count while the event got rowdy. You definitely would be disappointed you couldn't pull off such menial task. This software helps in solving that problem. It helps in creating a visual counter to help automate the task of having to count by yourself. Apart from the usability, functionality and ability to accomodate large numbers it also has a very fantastic UI